continued working on workflow operations

+++ src/jack/rm/workflow/RomConsolidator.java
@@ -11,28 +11,36 @@ import jack.rm.data.rom.Rom;
 public class RomConsolidator extends Dumper<RomHandle>
 {
   Path destination;
+  boolean overwrite;
+
   
   public RomConsolidator(Path path)
   {
     this.destination = path;
+    this.overwrite = false;
   }
   
   public void accept(RomHandle handle)
   {
     try
     {
-      if (handle.getBuffer() == null)
-        handle.prepareBuffer();
-      
       Rom rom = handle.getRom();
       Path basePath = destination.resolve(handle.getDestPath());
       Path finalPath = basePath.resolve(rom.getTitle()+"."+rom.getSystem().exts[0]);
+      
+      if (!overwrite && Files.exists(finalPath))
+        return;
+      
+      if (handle.getBuffer() == null)
+        handle.prepareBuffer();
+      
       Files.createDirectories(basePath);
       Files.move(handle.getPath(), finalPath, StandardCopyOption.REPLACE_EXISTING);
       //Files.deleteIfExists(handle.getPath());
     }
     catch (Exception e)
     {
+      System.out.println("Error on "+handle.getRom().getPath().toString());
       e.printStackTrace();
     }
   }
